+++ src/startplanner/__init__.py
@@ -1,3 +1,3 @@
 """StartPlanner package."""
 
-__version__ = "0.8.6"
+__version__ = "0.8.7"